+++ tests/Generator/MakeValueObjectTest.php
@@ -6,7 +6,8 @@ use Illuminate\Support\Str;
 use Lunarstorm\LaravelDDD\Tests\Fixtures\Enums\Feature;
 
 it('can generate value objects', function ($domainPath, $domainRoot) {
-    Config::set('ddd.paths.domains', $domainPath);
+    Config::set('ddd.domain_path', $domainPath);
+    Config::set('ddd.domain_namespace', $domainRoot);
 
     $valueObjectName = Str::studly(fake()->word());
     $domain = Str::studly(fake()->word());
@@ -48,7 +49,7 @@ it('normalizes generated value object to pascal case', function ($given, $normal
     $domain = Str::studly(fake()->word());
 
     $expectedPath = base_path(implode('/', [
-        config('ddd.paths.domains'),
+        config('ddd.domain_path'),
         $domain,
         config('ddd.namespaces.value_objects'),
         "{$normalized}.php",